+++ assembly/src/assembler/mast_forest_builder.rs
@@ -7,7 +7,7 @@ use core::ops::{Index, IndexMut};
 
 use miette::{IntoDiagnostic, Report};
 use vm_core::{
-    Decorator, DecoratorList, Operation,
+    Decorator, DecoratorList, Felt, Operation,
     crypto::hash::RpoDigest,
     mast::{
         DecoratorFingerprint, DecoratorId, MastForest, MastNode, MastNodeFingerprint, MastNodeId,
@@ -520,6 +520,12 @@ impl MastForestBuilder {
     fn fingerprint_for_node(&self, node: &MastNode) -> MastNodeFingerprint {
         MastNodeFingerprint::from_mast_node(&self.mast_forest, &self.hash_by_node_id, node)
             .expect("hash_by_node_id should contain the fingerprints of all children of `node`")
+    }
+
+    /// Registers an error message in the MAST Forest and returns the
+    /// corresponding error code as a Felt.
+    pub fn register_error(&mut self, msg: Arc<str>) -> Felt {
+        self.mast_forest.register_error(msg)
     }
 }